Fix needless collect warning
Clippy didn't like the fact that the Vec<QueryItem> was called query
along with the passed parameter named the same. Suggest fix was to
double up on .map(Into:into) in the variable definition. Running again
would have a suggested fix to double up on both of the .map(Into::into).